refactor: Move inventory types to models.inventory
