+++ common/src/main/java/com/wynntils/functions/InventoryFunctions.java
@@ -9,8 +9,8 @@ import com.wynntils.core.consumers.functions.Function;
 import com.wynntils.core.consumers.functions.arguments.FunctionArguments;
 import com.wynntils.core.text.PartStyle;
 import com.wynntils.core.text.StyledText;
-import com.wynntils.models.containers.type.InventoryAccessory;
-import com.wynntils.models.containers.type.InventoryArmor;
+import com.wynntils.models.inventory.type.InventoryAccessory;
+import com.wynntils.models.inventory.type.InventoryArmor;
 import com.wynntils.models.items.WynnItem;
 import com.wynntils.models.items.properties.DurableItemProperty;
 import com.wynntils.models.stats.type.ShinyStat;
