+++ common/src/main/java/com/wynntils/models/inventory/type/InventoryAccessory.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2023-2024.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.models.containers.type;
+package com.wynntils.models.inventory.type;
 
 import java.util.Locale;
